nonebot_plugin_jmcomic/pyproject.toml
[project]
name = "nonebot-plugin-jmcomic"
version = "0.1.6"
description = "下载禁漫本子并发送"
authors = [
    {name = "zhulinyv",email = "zhulinyv2005@outlook.com"}
]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_jmcomic"}]


[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.2.1"
jmcomic = "^2.5.32"
nonebot-adapter-onebot = "^2.2.3"
img2pdf = "=0.6.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
